Added SPDataAdapterBase<T extends ISPDataAdapterBaseConfiguration>.getMappedProjectProperties
