+++ SharePointFramework/shared-library/src/data/SPDataAdapterBase/types.ts
@@ -28,3 +28,23 @@ export enum ProjectAdminPermission {
   ProjectStatusAdmin = 'f6b875ae-fdb4-4ceb-bc75-ed853c2a2b0e',
   ChildProjectsAdmin = '2281c92a-f5ff-4d99-8814-e7b2f33d1ac9'
 }
+
+export type GetMappedProjectPropertiesOptions = {
+  /**
+   * Wrap multi values in results array.
+   */
+  wrapMultiValuesInResultsArray?: boolean
+
+  /**
+   * Use SharePoint taxonomy hidden fields to set the value for taxonomy fields.
+   * Default is using the custom text fields. 
+   * 
+   * @note `targetListName` must be set for this to work.
+   */
+  useSharePointTaxonomyHiddenFields?: boolean
+
+  /**
+   * Target list name to map the project properties for.
+   */
+  targetListName?: string
+}
